update userservice and finish controller

+++ src/main/java/com/example/springbootstarter/Services/Impl/UserServiceImpl.java
@@ -1,6 +1,8 @@
 package com.example.springbootstarter.Services.Impl;
 
 import com.example.springbootstarter.DTOS.CreateUserDTO;
+import com.example.springbootstarter.DTOS.UpdateUserDto;
+import com.example.springbootstarter.DTOS.UserGetDto;
 import com.example.springbootstarter.Repositories.UserRepository;
 import com.example.springbootstarter.Models.User;
 import com.example.springbootstarter.Services.UserService;
@@ -16,27 +18,55 @@ public class UserServiceImpl implements UserService {
     @Autowired
     private UserRepository userRepository;
 
+
     @Override
-    public User AddUser(CreateUserDTO dto) {
+    public UserGetDto AddUser(CreateUserDTO dto) {
         User mappedUser = dto.FromDto(dto);
         userRepository.save(mappedUser);
-        return mappedUser;
+        return mappedUser.ToDto();
     }
 
     @Override
-    public Optional<User> GetUser(long id) {
-        return userRepository.findById(id);
+    public Optional<UserGetDto> GetUserById(long id) {
+        Optional<User> user = userRepository.findById(id);
+       return user.map(User::ToDto);
     }
 
     @Override
-    public List<User> GetUsers() {
-        ArrayList<User> users = new ArrayList<>();
-        userRepository.findAll().forEach(users::add);
+    public Optional<UserGetDto> GetUserByEmail(String email) {
+        Optional<User> user = userRepository.findByEmail(email);
+        return user.isEmpty() ? Optional.empty() : Optional.of(user.get().ToDto());
+    }
+
+    @Override
+    public List<UserGetDto> GetUsers() {
+        ArrayList<UserGetDto> users = new ArrayList<>();
+        userRepository.findAll().forEach(user -> {users.add(user.ToDto());});
         return users;
     }
 
     @Override
-    public void DeleteUser(long id) {
+    public Optional<UserGetDto> UpdateUser(UpdateUserDto dto) {
+        Optional<User> user = userRepository.findById(dto.Id);
+
+        if(user.isEmpty()){
+            return Optional.empty();
+        }
+        User userToUpdate = user.get();
+        userToUpdate.setName(dto.Name);
+        userToUpdate.setEmail(dto.Email);
+
+        userRepository.save(userToUpdate);
+        return Optional.of(userToUpdate.ToDto());
+    }
+
+    @Override
+    public Optional<Long> DeleteUser(long id) {
+        Optional<User> user = userRepository.findById(id);
+        if(user.isEmpty()){
+            return Optional.of(id);
+        }
         userRepository.deleteById(id);
+        return Optional.empty();
     }
 }
